// src/app/api/Prix.ts
import { ProfessionalLightDTO } from "./professionalLightDTO";

export class Prix{

    id: number;
    professional: ProfessionalLightDTO;
    prestation: string;
    prestation1: string;
    prestation2: string;
    prestation3: string;  
    prix: string;
    prix1: string;
    prix2: string;
    prix3: string;
}